+++ test/LibSSHTests.jl
@@ -561,6 +561,22 @@ end
             # Test setting environment variables
             cmd = setenv(`echo \$foo`, "foo" => "bar")
             @test readchomp(cmd, session) == "bar"
+
+            # Test command failure. We redirect error output to devnull to hide
+            # the errors displayed by the errormonitor() task.
+            try
+                redirect_stderr(devnull) do
+                    run(`foo`, session)
+                end
+            catch ex
+                @test ex isa TaskFailedException
+                @test current_exceptions(ex.task)[1][1] isa ssh.SshProcessFailedException
+            end
+
+            # Test passing a String instead of a Cmd
+            mktempdir() do tmpdir
+                @test readchomp("cd $(tmpdir) && pwd", session) == tmpdir
+            end
         end
     end